Refactor proptest-support for fenris-geometry
